src/test/kotlin/org/example/e3FxGaming/adventOfCode2022Kotlin/Day08Test.kt
package org.example.e3FxGaming.adventOfCode2022Kotlin

import org.junit.jupiter.api.Assertions.*
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.MethodSource
import java.io.Reader
import java.util.stream.Stream

class Day08Test {
    companion object {
        @JvmStatic
        fun argumentsProvider(): Stream<Arguments> = Stream.of(
            Arguments.of(Day08Test::class.java.getResource("/input/day08.txt").toReader(), 21, 8)
        )
    }

    @ParameterizedTest
    @MethodSource("argumentsProvider")
    fun day08(inputReader: Reader, part1: Int, part2: Int) {
        Day08(inputReader).let {
            assertEquals(part1, it.part1(), "Part 1")
            assertEquals(part2, it.part2(), "Part 2")
        }
    }
}